Support v0.3 schema report format in LLO
- Support types of stream value other than big int
- Support multiple different methods of aggregation
- Closes MERC-3525
