+++ llo/channel_definitions.go
@@ -2,6 +2,7 @@ package llo
 
 import (
 	"fmt"
+	"sort"
 
 	llotypes "github.com/smartcontractkit/chainlink-common/pkg/types/llo"
 )
@@ -26,4 +27,29 @@ func VerifyChannelDefinitions(channelDefs llotypes.ChannelDefinitions) error {
 		return fmt.Errorf("too many unique stream IDs, got: %d/%d", len(uniqueStreamIDs), MaxObservationStreamValuesLength)
 	}
 	return nil
+}
+
+func subtractChannelDefinitions(minuend llotypes.ChannelDefinitions, subtrahend llotypes.ChannelDefinitions, limit int) llotypes.ChannelDefinitions {
+	differenceList := []ChannelDefinitionWithID{}
+	for channelID, channelDefinition := range minuend {
+		if _, ok := subtrahend[channelID]; !ok {
+			differenceList = append(differenceList, ChannelDefinitionWithID{channelDefinition, channelID})
+		}
+	}
+
+	// Sort so we return deterministic result
+	sort.Slice(differenceList, func(i, j int) bool {
+		return differenceList[i].ChannelID < differenceList[j].ChannelID
+	})
+
+	if len(differenceList) > limit {
+		differenceList = differenceList[:limit]
+	}
+
+	difference := llotypes.ChannelDefinitions{}
+	for _, defWithID := range differenceList {
+		difference[defWithID.ChannelID] = defWithID.ChannelDefinition
+	}
+
+	return difference
 }
